+++ scanner.js
@@ -79,8 +79,11 @@ export async function createBarcodeScanner(callback = console.log, {
 			: controller.signal;
 
 		const scanner = new BarcodeDetector({ formats });
-		const wakeLock = 'wakeLock' in navigator ? await navigator.wakeLock.request('screen').catch(() => undefined) : undefined;
 		const video = document.createElement('video');
+		const wakeLock = 'wakeLock' in navigator
+			? await navigator.wakeLock.request('screen').catch(() => undefined)
+			: undefined;
+
 		const stream = await navigator.mediaDevices.getUserMedia({
 			audio: false,
 			video: {
@@ -92,13 +95,16 @@ export async function createBarcodeScanner(callback = console.log, {
 		});
 
 		const [track] = stream.getVideoTracks();
+		const capture = globalThis?.ImageCapture?.prototype?.grabFrame instanceof Function
+			? new ImageCapture(track)
+			: undefined;
 
 		video.srcObject = stream;
 		video.play();
 
 		async function drawFrame() {
 			try {
-				const results = await scanner.detect(video).catch(err => {
+				const results = await scanner.detect(typeof capture === 'undefined' ? video : await capture.grabFrame()).catch(err => {
 					reportError(err);
 					return [];
 				});
@@ -127,8 +133,10 @@ export async function createBarcodeScanner(callback = console.log, {
 		}, { once: true, signal: sig });
 
 		video.addEventListener('error', () => {
-			loadController.abort();
-			controller.abort(new DOMException('Error loading video stream.'));
+			const err = new DOMException('Error loading video stream.');
+			loadController.abort(err);
+			controller.abort(err);
+			reject(err);
 		}, { once: true, signal: sig });
 
 		sig.addEventListener('abort', async ({ target }) => {